+++ core/components/rememberthis/lexicon/de/default.inc.php
@@ -4,22 +4,18 @@ $_lang['rememberthis_desc'] = "Remember an id and the correspondenting resource 
 $_lang['rememberthis.add'] = 'Merken';
 $_lang['rememberthis.delete'] = 'Entfernen';
 $_lang['rememberthis.noresultstext'] = 'Die Liste ist leer';
-$_lang['rememberthis.rememberhead'] = 'Merkliste';
 
 $_lang['setting_rememberthis.rowTpl'] = 'Row Template';
 $_lang['setting_rememberthis.rowTpl_desc'] = 'Row templage for the list output.';
 
 $_lang['setting_rememberthis.outerTpl'] = 'Outer Template';
-$_lang['setting_rememberthis.outerTpl_desc'] = 'Outer templage for the list output.';
+$_lang['setting_rememberthis.outerTpl_desc'] = 'Outer templage for the list output, if the list is not empty.';
 
 $_lang['setting_rememberthis.addTpl'] = 'Add Template';
 $_lang['setting_rememberthis.addTpl_desc'] = 'Template for the add link.';
 
 $_lang['setting_rememberthis.noResultsTpl'] = 'No Results Template';
 $_lang['setting_rememberthis.noResultsTpl_desc'] = 'Template that is displayed, if the list is empty.';
-
-$_lang['setting_rememberthis.rememberhead'] = 'Head Template';
-$_lang['setting_rememberthis.rememberhead_desc'] = 'Template that is displayed as head, if the list is not empty.';
 
 $_lang['setting_rememberthis.itemTitleTpl'] = 'Item Title Template';
 $_lang['setting_rememberthis.itemTitleTpl_desc'] = 'Template for one list item.';